requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
motor==3.4.0
pymongo==4.7.2
pymongoarrow==1.3.1
python-dotenv==1.0.1
python-decouple==3.8
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
firebase-admin
pydantic[email]